Restructure the Echo Funlet
I extracted and exported the echo() function which handles
the core logic synchronously, separated from the complexity
of the asynchronous handler.

This allowed me to simplify the units tests, which now run three
simple synchronous tests and a single complex asynchronous text;
and only the latter needs to be skipped until the test harness
is completed with Twilio.Response.