Complete implementation of Simple Message Funlet with handler
I also updated the getter for messages to clarify that it returns
multiple messages in the general case. This is reflected in the name
of the getter function, which is now `getMessages()`, as well as in the
constant for the default value, now `MY_MESSAGES`, and the README
descriptions. The name of the GET/POST parameter remains `Message`
(singular) to preserver compatibility with the Twimlet.