Refine showRedirect() to display either URL or indented body
While the Twimlets used HTTP redirects most of the time, it is simpler
in the corresponding Funlets to use a TwiML <Redirect/> instead.

I rewrote showRedirect() to make a full HTTP query (not limited to headers)
with an extra status which prints the redirect URL. When the response is
empty, it starts directly with this extra status, which is the only line
printed. Otherwise, I discard this status and indent the response instead.