Remove distinction between GET and POST Event parameters
Funlets are intended to be set up as webhooks/handlers for Twilio Voice
events. Custom POST parameters are unlikely to be provided in this context.
It is also unlikely that these custom POST parameters would be preserved
in a redirect back to the script through <Redirect/> or <Gather/>.

Therefore, I have simplified the description of expected parameters,
removing special cases where `POST` parameters were allowed a more
developed representation than `GET` parameters.