Handle TwiML <Redirect/> when document does not end with a newline
I previously expected the XML document to end with a newline,
which pushed the curl status to its own line. I found that this
was not the case after actual deployment of a Twilio Function. I
have now changed the `sed` command to remove the status line from
the end of the document, not necessarily starting from a newline.