Add HumanCheck=1 for integration tests with Whisper Twimlet
In the Find Me Funlet, the HumanCheck is set to true by default,
to reflect the fact that the Find Me Funlet sets this flag to true
in the params for the call to the Whisper script.

Adding HumanCheck=1 systematically allows to get the corresponding
output for the Whisper Twimlet when the check is required.