Treat any value of HumanCheck flag except "false" as true
First, I only treated "true" as true and everything else as false.
This was not consistent with the behavior of PHP boolean parameters,
which are mostly true.

Then, I considered that the flag was always true when any value was set.
This was consistent with the PHP implementation of the script, but it
prevented from overriding a default value that happens to be true,
since setting any other value can only make the flag true, and not false.

In order to allow forcing this flag to false, without handling all the
complexity of PHP truth values, I decided to add just one false value,
the string "false".

If needed for enhanced compatibility, a second false value may be added
at a later point, the string "0".