Reworks the sms-notifications phone number UI
The phone number UI for the sms-notifications function template now
includes an Add button that replicates the functionality of the Enter
button for the phone number input. It also now automatically adds the
current value of the phone number input (if any) to the recipient list
when Send is pressed.