Adds an example contact form
I've found that a question that comes up often is how to create a contact form with SendGrid. The difficulty is that you would expect to send the email from the email address of the person submitting the form. To ensure that SendGrid only sends reputable emails, you normally have to verify your email address or authenticate your domain first. So, this contact form sends the email from a verified/authenticated email address and places the submitter's email in the reply-to field so when you reply to the email it goes back to the original submitter.